Only tested with Linux 4.13 and FreeBSD 10.3
Let's say theoretically I--I mean, you--are making daily exports of the configuration of my--I mean your--switches, and saving each of them in a folder named 
switch_configs/$(date +%Y-%m-%d_%H-%M)/$switchname.txt
We can't just test the root directory to see when it was modified, right? That would only tell you that any attempt at a config export was made, not which config nor how successfully it was exported (and in the first iteration of the switch backup script I won't publish, the last config to be saved would always be 0 bytes in size).

This will search through a possibly segfaultingly high number of subdirectories to return every file that matches the parameters specified in our Nagios host config, which obviously resides elsewhere.

Parameters must not contain spaces. If you use spaces in your folder structure then fix your shame, patch my script, or make a symlink.